requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
sqlalchemy>=2.0
psycopg2-binary>=2.9
pydantic>=2.10
pydantic-settings>=2.6
python-multipart>=0.0.9
python-dotenv>=1.0
email-validator>=2.0
httpx>=0.27
pdfplumber>=0.10
openpyxl>=3.1
anthropic>=0.50
gspread>=6.0
google-auth>=2.0
click>=8.0
